move types to index
